+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * improve mobile support
 * ignore empty command in Pipe extension [#984](https://github.com/jcubic/jquery.terminal/issues/984)
 * fix processing Hex HTML entities [#992](https://github.com/jcubic/jquery.terminal/issues/992)
+* fix returning Terminal instance from interpreter [#994](https://github.com/jcubic/jquery.terminal/issues/994)
 
 ## 2.44.1
 ### Bugfix